Move PerKeyCombineFnRunners to runners/core.
